+++ src/components/profile/ColorGenerator.jsx
