+++ src/pages/Profile.jsx
@@ -1,12 +1,12 @@
 import { useState, useEffect } from "react";
 import { useNavigate } from "react-router-dom";
-import { useAuth } from "../hooks/useAuth";
+import { useAuth } from "../hooks/useAuth.js";
 import { verifySession, updateUser, deleteUser } from "../api/user.js";
 
 import Container from "../ui/Container.jsx";
-import ProfileForm from "../components/Profile/ProfileForm.jsx";
+import ProfileForm from "../components/profile/ProfileForm.jsx";
 import ModalConfirm from "../ui/ModalConfirm.jsx";
-import ModalChangePassword from "../components/Profile/ModalChangePassword.jsx";
+import ModalChangePassword from "../components/profile/ModalChangePassword.jsx";
 import Button from "../ui/Button.jsx";
 
 export default function Profile() {
