+++ src/pages/Quiz.jsx
@@ -2,8 +2,8 @@ import { useParams, useNavigate } from "react-router-dom";
 import { saveResult, getResultById } from "../api/results.js";
 import { getQuizById } from "../api/quizzes.js";
 import { useState, useEffect } from "react";
-import { useAuth } from "../hooks/useAuth";
-import Question from "../components/Quiz/Question.jsx";
+import { useAuth } from "../hooks/useAuth.js";
+import Question from "../components/quiz/Question.jsx";
 import Button from "../ui/Button.jsx";
 import Container from "../ui/Container.jsx";
 import ModalConfirm from "../ui/ModalConfirm.jsx";
