+++ src/pages/Results.jsx
@@ -1,10 +1,10 @@
 import { useState, useEffect, useCallback } from "react";
 import { useNavigate, Link } from "react-router-dom";
 import { getResults } from "../api/results.js";
-import { useAuth } from "../hooks/useAuth";
-import { useDebounce } from "../hooks/useDebounce";
-import Grid from "../components/Home/Grid.jsx";
-import ToolBar from "../components/Home/ToolBar.jsx";
+import { useAuth } from "../hooks/useAuth.js";
+import { useDebounce } from "../hooks/useDebounce.js";
+import Grid from "../components/home/Grid.jsx";
+import ToolBar from "../components/home/ToolBar.jsx";
 
 const ITEMS_PER_PAGE = 36;
